better initial mainwindow placement

+++ fundo/core/Search/SearchResult.cs
@@ -5,15 +5,15 @@ namespace fundo.core.Search
     internal class SearchResult
     {
         private String fileName;
-        private int fileSize;
+        private long fileSize;
         private DateTime fileDate;
 
         public string FileName { get => fileName; set => fileName = value; }
-        public int FileSize { get => fileSize; set => fileSize = value; }
+        public long FileSize { get => fileSize; set => fileSize = value; }
         public DateTime FileDate { get => fileDate; set => fileDate = value; }
 
 
-        public SearchResult(String fileName, int fileSize, DateTime fileDate)
+        public SearchResult(String fileName, long fileSize, DateTime fileDate)
         {
             this.fileName = fileName;
             this.fileSize = fileSize;
